infra-as-code/modules/export-to-bq/function-source-code/requirements.txt
functions-framework==3.*
requests
google-auth
google-cloud-bigquery